// modules/users/user-usage.entity.d.ts
import { User } from './user.entity';
export declare class UserUsage {
    id: string;
    user: User;
    user_id: string;
    session_id: string;
    action: string;
    entity_type: string;
    entity_id: string;
    latency_ms: number;
    created_at: Date;
}
//# sourceMappingURL=user-usage.entity.d.ts.map